backend/requirements.txt
# 배짱이 v1.0 — 백엔드 의존성
# Python 3.11+ 권장

fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# DB: SQLite + 비동기
sqlalchemy[asyncio]==2.0.27
aiosqlite==0.19.0

# 인증/보안
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
cryptography==42.0.4
google-auth>=2.0.0

# 검증
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP/WebSocket
httpx==0.27.0
websockets==12.0

# 업비트 API
pyupbit==0.2.33

# 로깅
loguru==0.7.2

# FCM (Firebase Cloud Messaging)
firebase-admin>=6.0.0
